fetch yearly data for resorts
The notebook will now, each time it is run, attempt to find
a year of missing data in hopsworks for closed or open resorts. If it finds
such a year, it will
* create weather data for that year
* upload the weather data to hopsworks
* exit

In this way, the notebook can be run in a pipeline-like manner, for
instance fetching a new year of data for every pipeline.